+++ Common/src/main/java/fuzs/puzzleslib/common/api/util/v1/ValueSerializationHelper.java
@@ -49,6 +49,7 @@ public final class ValueSerializationHelper {
             } else {
                 valueOutput = TagValueOutput.createWithoutContext(scopedCollector);
             }
+
             valueOutputConsumer.accept(valueOutput);
             return valueOutput.buildResult();
         }