+++ Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/event/SpawnReasonMob.java
@@ -18,6 +18,7 @@ import org.jspecify.annotations.Nullable;
  * {@link net.minecraft.world.entity.Mob#finalizeSpawn(ServerLevelAccessor, DifficultyInstance, EntitySpawnReason,
  * SpawnGroupData)} was overridden without calling {@code super}.
  */
+@Deprecated
 public interface SpawnReasonMob {
     /**
      * Use {@link Enum#name()} for backward compatibility.